Move authenticate/public to autnenticate/public/checkout, which reflects what it will become
